requirements.txt
grpcio==1.32.0
grpcio-tools==1.32.0
protobuf==3.13.0
python-dotenv==0.14.0
six==1.15.0
